Extract IR operations to ir_ops.rs module
Move IR operation compilation from ir_to_ebpf.rs to a dedicated ir_ops.rs
module with an IrOps extension trait:

- compile_load_literal: literals (int, bool, string, record)
- compile_move: register moves
- compile_binary_op: math, bitwise, and comparison operations
- compile_comparison: comparison result generation
- compile_return: program exit
- compile_store_variable/compile_load_variable: variable handling
- compile_not: logical negation

The ir_to_ebpf.rs file is now reduced from ~1,154 to 927 lines, making
the compiler more maintainable. The modularization effort has reduced
the original 2,293 line monolith to focused modules:
- ir_to_ebpf.rs: 927 lines (core compiler, control flow, dispatch)
- register_alloc.rs: 260 lines (register allocation with spilling)
- ir_ops.rs: 301 lines (IR operation compilation)
- helpers/: ~1,100 lines across 5 files (BPF helper compilation)
